Solaris 10: Fix disk partitioning in JumpStart
Target `any` disk when partitioning. Otherwise things are very sensitive to the
order in which disks are attached to the VM (different packer builders attach
devices in different orders) and whether SATA or IDE interfaces are in use.

Targetting `any` disk just does the right thing as long as only one virtual
disk is in use (which is currently the only option available).